Move concatenation code into its own header
