+++ Makefile.am
@@ -33,6 +33,7 @@ lepton_SOURCES = \
    src/lepton/socket_serve.cc \
    src/lepton/socket_serve.hh \
    src/lepton/jpgcoder.cc \
+   src/lepton/concat.cc \
    src/lepton/main.cc \
    src/lepton/validation.hh \
    src/lepton/validation.cc \
